use the 'int' module in the 'primes' module

+++ src/primes.rs
@@ -4,6 +4,7 @@ extern crate itertools;
 use self::itertools::Itertools;
 
 use std::collections::HashSet;
+use super::int::Sqrt;
 
 // base struct for a prime numbers iterator, that stores encountered primes
 // in order to speed-up discovery of subsequent primes.
@@ -58,7 +59,7 @@ pub fn is_prime(n: u64) -> bool {
     if n == 2 { return true; }
     if n % 2 == 0 { return false; }
 
-    let lim = (n as f64).sqrt() as u64 + 1;
+    let lim = n.sqrt() + 1;
     for x in (3..lim).step(2) {
         if n % x == 0 {
             return false;
@@ -81,7 +82,7 @@ pub fn is_prime_with_set(n: u64, set: &HashSet<u64>) -> bool {
 pub fn divisors(n: u64) -> Vec<u64> {
     if n == 1 { return vec![1]; }
 
-    let mut s = (n as f64).sqrt() as u64;
+    let mut s = n.sqrt();
     let square = s * s == n;
     if !square { s += 1 }
 
@@ -107,7 +108,7 @@ fn next_prime_factor(n: u64) -> u64 {
     if n == 1 { return 0; }
     if n % 2 == 0 { return 2; }
 
-    let lim = (n as f64).sqrt() as u64 + 1;
+    let lim = n.sqrt() + 1;
     for x in (3..lim).step(2) {
         if n % x == 0 {
             return x;
